1634: Little project fixes * Maven Shade Plugin upgraded to version 3.2.0 * Added httpclient.version and bower.chart.js.version variables to `pom.xml` * Menas: Same origin headers * `ConfigReaderSuite` moved to correct directory
