+++ menas/pom.xml
@@ -47,7 +47,7 @@
         <dependency>
             <groupId>org.apache.httpcomponents</groupId>
             <artifactId>httpclient</artifactId>
-            <version>4.4.1</version>
+            <version>${httpclient.version}</version>
         </dependency>
         <!-- Spark -->
         <dependency>
@@ -251,7 +251,7 @@
         <dependency>
             <groupId>org.webjars.bower</groupId>
             <artifactId>chart.js</artifactId>
-            <version>2.7.3</version>
+            <version>${bower.chart.js.version}</version>
         </dependency>
         <!-- Data formats -->
         <dependency>
