+++ menas/src/main/scala/za/co/absa/enceladus/menas/WebSecurityConfig.scala
@@ -72,6 +72,9 @@ class WebSecurityConfig @Autowired()(beanFactory: BeanFactory,
         .and()
           .addFilterBefore(kerberosFilter, classOf[UsernamePasswordAuthenticationFilter])
           .addFilterAfter(jwtAuthFilter, classOf[SpnegoAuthenticationProcessingFilter])
+        .headers()
+          .frameOptions()
+          .sameOrigin()
     }
 
     @Bean
