+++ pom.xml
@@ -136,7 +136,7 @@
         <maven.surefire.version>2.18.1</maven.surefire.version>
         <maven.scoverage.version>1.3.0</maven.scoverage.version>
         <maven.scala.version>3.2.0</maven.scala.version>
-        <maven.shade.version>2.3</maven.shade.version>
+        <maven.shade.version>3.2.0</maven.shade.version>
         <maven.sources.version>3.0.1</maven.sources.version>
         <maven.war.version>2.2</maven.war.version>
         <maven.ant.plugin.version>1.8</maven.ant.plugin.version>
@@ -191,6 +191,8 @@
         <cronstrue.version>1.79.0</cronstrue.version>
         <wiremock.version>2.26.3</wiremock.version>
         <guava.version>27.0.1-jre</guava.version>
+        <httpclient.version>4.4.1</httpclient.version>
+        <bower.chart.js.version>2.7.3</bower.chart.js.version>
         <!--other properties-->
         <skip.integration.tests>true</skip.integration.tests>
         <quiet.scalastyle.check>true</quiet.scalastyle.check>
