+++ spark-jobs/src/test/scala/za/co/absa/enceladus/conformance/interpreter/InterpreterSuite.scala
@@ -15,8 +15,6 @@
 
 package za.co.absa.enceladus.conformance.interpreter
 
-import org.json4s._
-import org.json4s.native.JsonParser._
 import org.mockito.Mockito.{mock, when => mockWhen}
 import org.scalatest.{BeforeAndAfterAll, FunSuite}
 import za.co.absa.atum.model.ControlMeasure
@@ -24,8 +22,10 @@ import za.co.absa.enceladus.conformance.config.ConformanceConfig
 import za.co.absa.enceladus.conformance.datasource.DataSource
 import za.co.absa.enceladus.dao.MenasDAO
 import za.co.absa.enceladus.conformance.samples._
-import za.co.absa.enceladus.utils.fs.FileReader
 import za.co.absa.enceladus.utils.testUtils.{LoggerTestBase, SparkTestBase}
+import org.json4s._
+import org.json4s.jackson._
+import za.co.absa.enceladus.utils.fs.FileReader
 
 class InterpreterSuite extends FunSuite with SparkTestBase with BeforeAndAfterAll with LoggerTestBase {
 
@@ -84,7 +84,7 @@ class InterpreterSuite extends FunSuite with SparkTestBase with BeforeAndAfterAl
 
     implicit val formats: DefaultFormats.type = DefaultFormats
 
-    val checkpoints = parse(infoFile).extract[ControlMeasure].checkpoints
+    val checkpoints = parseJson(infoFile).extract[ControlMeasure].checkpoints
 
     assertResult(expected)(data)
     // test drop
@@ -148,7 +148,7 @@ class InterpreterSuite extends FunSuite with SparkTestBase with BeforeAndAfterAl
 
     implicit val formats: DefaultFormats.type = DefaultFormats
 
-    val checkpoints = parse(infoFile).extract[ControlMeasure].checkpoints
+    val checkpoints = parseJson(infoFile).extract[ControlMeasure].checkpoints
 
     if (data != expected) {
       logger.error("EXPECTED:")
