+++ utils/src/test/scala/za/co/absa/enceladus/utils/config/ConfigReaderSuite.scala
@@ -13,11 +13,10 @@
  * limitations under the License.
  */
 
-package za.co.absa.enceladus.utils.general
+package za.co.absa.enceladus.utils.config
 
 import com.typesafe.config.ConfigFactory
 import org.scalatest.WordSpec
-import za.co.absa.enceladus.utils.config.ConfigReader
 
 class ConfigReaderSuite extends WordSpec {
   private val config = ConfigFactory.parseString(
